requirements.txt
torch>=1.1.0
torchvision>=0.3.0
opencv>=3.4.0
Pillow>=6.2.0
pyaml
yacs
tqdm
coloredlogs